% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XeniumSpatialAnalysisWorkshop.R
\docType{data}
\name{ref_wu}
\alias{ref_wu}
\title{Pseudo-bulked single-cell breast cancer reference dataset}
\format{
A SingleCellExperiment object.
}
\usage{
ref_wu
}
\description{
A pseudo-bulked single-cell RNA-seq breast cancer dataset from the study by Wu et al., Nat Genetics, 2021. The original dataset has been subsetted to the genes present in the 10x Xenium dataset.
}
\references{
Wu SZ, Al-Eryani G, Roden DL, Junankar S, Harvey K, Andersson A, Thennavan A, Wang C, Torpy JR, Bartonicek N, Wang T. A single-cell and spatially resolved atlas of human breast cancers. Nature genetics. 2021 Sep;53(9):1334-47. https://www.weizmann.ac.il/sites/3CA/breast
}
\keyword{datasets}
